+++ src/interfaces/IPairFactory.sol
@@ -1,7 +1,7 @@
 pragma solidity ^0.8.15;
 
 interface IPairFactory {
-    function allPairsLength() external view returns (uint);
+    function allPairsLength() external view returns (uint256);
     function isPair(address pair) external view returns (bool);
     function pairCodeHash() external pure returns (bytes32);
     function getPair(address tokenA, address token, bool stable) external view returns (address);
